# Use Table lookup for HPACK decoder