@normanmaurer 2019-06-29
@netty-bot test this please