@normanmaurer 2019-07-01
@carl-mastrangelo i think we should add the license header of hyper and also add it to the NOTICE file.

@Lukasa can you please have a look as well ?