@carl-mastrangelo 2019-07-01
@normanmaurer the code was modified from Python hyper, but it claims the code came from nginx.   Not sure who's copyright to include.