@normanmaurer 2019-07-01
@netty-bot test this please