@normanmaurer 2019-07-01
/cc @vkostyukov @mosesn as we spoked about http2 header / hpack perf in the past