@normanmaurer 2019-07-02
Also please note that this is reproducable every time.