@normanmaurer 2019-07-02
That said the code is a lot nicer as it is imho so we should decide if we want to keep it clean or want to get the last % perf :) /cc @ejona86