@carl-mastrangelo 2019-07-02
@normanmaurer for performance, I would like the speed boost, I just wasn't able to reproduce it on my machine.   I'd be happy to review the byte processor change after this.