@normanmaurer 2019-07-02
@netty-bot test this please