@ejona86 2019-07-02
I think I'd be fine if you swapped to the processor. The method is pretty small and you can show efficiency gains of ~3%.