@carl-mastrangelo 2019-07-02
@ejona86 

I tried out the tables in chttp2, and the performance was dramatically worse, about half.   Looking at perfasm, I am guessing it's due to the 4x indirection of the table lookup, based on the hottest region analysis.  

I would hazard a guess that my implementation is competitive with the C version, despite using a more bloated table.