@normanmaurer 2019-07-03
Here we go: https://github.com/netty/netty/pull/9317